+++ src/main/java/nextstep/app/web/RoomEscapeWebApplication.java
@@ -1,0 +1,11 @@
+package nextstep.app.web;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RoomEscapeWebApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(RoomEscapeWebApplication.class, args);
+    }
+}
